A working editor, nerf1 style.
This version is the result of some opportunistic "scavenging" of nerf0. That is: copy/paste as needed.
The baseline in nerf0 for this action was 757e9c595a65.

The main difference with nerf0 is the different Clef (which was introduced already in the previous commit). That, in
turn, has some consequences:

In nerf0 the notes `Insert` and `Extend` take a full history in the form of a nout_hash. In nerf1 their equivalents take
only a single note.

The approach in nerf0 had as a consquence that all communications over channels were in the form of nout_hashes, and
that a lot of explicit nout_hash handling was going on. Basically: because we depended on explicit handling of
nout_hashes in _some_ cases, we made this the interface in _all_ cases.

Similarly: in nerf0 the expectation in channel-communications was that you could send any nout_hash, and that this would
be picked up in the form of a Replace on the other end. We no longer have such an expectation, because we've basically
switched to a extend-only version of Replace (which is called Extend to reflect this).

In nerf1 we no longer have this requirement, so we've moved away from nout_hashes as the means of communication over
channels. This also means: we've gotten rid of the Possibility/Actuality duality, as the nerf1 clef's notes are
self-contained. The resulting simplification of dsn/s_expr/utils.py can be seen as a validation of this approach.

The other big difference with nerf0 is the display of history, which is now in the form of the ELS paper, i.e.
s-expressions rather than some rather cryptic bunch of symbols.

The following things were explictly not copied over yet ('scavenged'):

* the evaluator, a.k.a. the interpreter (PoC)
* form_analysis (PoC, currently on hold)
* historiography (not needed in an extend-only context)
* weaving (PoC only)
* fonts (broken anyway)
* tools for debugging & optimization (we'll scavenge them as needed)
* from_python.py (perhaps later, as needed)

Here are some other differences with nerf0:

* Note.metadata => Note.score
* Note.error is not preserved (we'll scavenge as needed)
* Editing of history is not supported (it would imply: any-replace; if we reintroduce it it will be in a self-applicable way)
* s/t addressing in HistoryWidget was not preserved

This remains to be done:

* Implement the editor Actions: MoveSelectionChild, MoveSelectionSibbling, LeaveChildrenBehind, EncloseWithParent