(Breaking refactoring) trees broadcast t-address
In working towards animations that display what's actually going on on the
history-side when selecting a different element in the tree, a
refactoring was needed:

Previous to the refactoring, the tree would communicate a cursor-change
to the history-widget by sending the full new score to display (with
some additional indirection as well). When a deeper node in the tree was
selected, the score of that node would be sent to the history, and the
history would render it.

However, we want to make some "animations" on the history-widget
(visualisations of state transitions) that show how pre-cursor-change
and post-cursor-change relate. To do this, the history-widget needs to
know this relationship. That is: it needs to have access to the global
history as well as the information that only part of that global history
must be displayed.

To achieve that, we've split out the 2 pieces of information into 2
channels and associated send methods: one for the full score (of all
history) and one for the cursor changes; The former is connected in
editor.py, by recycling the child-channel creation mechanism; the latter
is using a much simplified version of broadcast_cursor_update.

The refactoring also removes quite some complexity that was put in place
with the specific goal of not sharing the t-address space from the tree
to the history widget. (i.e. it removes the aformementioned indirection)

Breaking refactoring, because the history no longer shows what's under the cursor.
We will "fix this soon"