Fix previous breakage: history-widget follows tree again
Broken in ae8b843b281e; in this fixing commit the histories now
calculate what their "local history" is, that is: the history that's
actually corresponding to the currently selected node on the tree.

Contains ample duplication, this is a simpe consequence of the fact that
I didn't bother to deduplicate communalities between history.py and
ic_history.py yet.

Some additional notes:

* Previous cursor-update was depending on the invalidation in
    update_score, i.e. it only worked by accident; fixed, we always
    reconstruct boxes and invalidate in `parent_cursor_update` now.

* Cursor channel updates always follow score updates. This ensures that
    the cursor can be resolved (if it's pointing to a newly created
    thing).