Proof of concept of "animations" for history.
WHAT ARE ANIMATIONS?

Animations build on the following feature (one that was already existing
pre-commit): whenever we select an s-expression in our "tree" (the structural
view on the right hand side of the window) we show the history of that
particular s-expression on the left. That is, whenever we change the cursor in
the tree, we switch what is shown in the history.

The animation that's new in this commit is the following: any part of history
that shows up both before and after this switch will "float" from its pre-switch
position to its post-switch position in a number of steps. The idea is to make
it visually more clear that there is a relationship between the histories at
different levels of the tree.

STATUS: POC

This is truly meant as a Proof of Concept; it's non-production-ready in at least
the following ways:

* Non-performant on anything but toy examples
* Violations of "Don't Repeat Yourself" / General hackiness

IDENTITY OF NOTES

The key idea in the animations is to float textures from some pre-switch to a
post-switch location. This hinges on the assumption that we have a shared
identity for the textures pre- and post-switch. E.g. to float some open-bracket
from one location to the next, we need to know which open-bracket we're talking
about (there are many).

The solution is to have some addressing scheme for the textures that is global
in the sense that it is shared between the pre-and post-switch environments.
Using this addressing scheme we can identify textures: same address means same
texture.

Such an addressing scheme is obtained in a number of steps: first, we annotate
the "global history" (the history of the whole tree) in such a way that we can
uniquely identify each note.

Second, we construct a tree by playing this global history of notes, annotated
with their global address.  We use the same mechanism for this as we did
pre-commit. The difference is: because the input Notes have now been annotated
with a global address, the scores as constructed at each sub-expression in the
resulting tree are now consisting of notes which have a global address. This
means that when we fetch the "local score" (the score to be rendered) we have
information about the global address of each note.

Finally, we make sure to keep the annotations around in each step of the
conversion to textures, as well as add conversion-specific information when
needed.  In particular, the tree of notes is first converted to an s-expr (to be
able to render in the style of the ELS'18 paper), and these s-expressions are
then converted to the actual textures with locations.

We need step-specific address information for each such step: when converting to
an s-expression, we annotate the elements that are specific to the fact that the
note is being rendered as an s-expression (i.e. the fact that the Note's fields
and its type, when converted to an s-expression, turn into particular
further s-expressions); when converting to textures similar further annotations
are necessary (e.g. a list-expression is rendered as 2 textures, one for each
bracket)

(The above hinges on the fact that history widgets have access to a history
that's greater than what they are showing _right now_; the work needed for that
was done in a previous commit, namely ae8b843b281e)

THE ANIMATION

The actual animation is rather straightforward: do a linear interpolation for
(source, target) for the attributes (x, y, alpha).

We set a clock at an interval (I've set 1/60, but I'm not actually getting this
_at all_ on my local machine).  Kivy will tell you how much time has actually
passed since the last tick. We then calculate the fraction dt/remaining_time.
This approach is automatically robust for missed frames (i.e. the missed frame
will not be rendered, but the total animation time and the position of the
texture at the next frame are unaffected)

ROAD NOT TAKEN: UTILIZING NonTerminalBox

The existing code for rendering had some support for rendering hierarchical
structures of "boxes", recursively at some offset each.

One motivation for this was the belief that at some point in the future such
structures could be used to cheaply calculate rerenderings, given a small
change. (The idea: such changes affect one box at each level of the hierarchy
only, making reuse of the other already rendered boxes possible)

At first glance, it seemed possible to reuse these hierarchies in the
animations, floating whole NonTerminalBoxes rather than individual textures.
However, such reuse hinges on the assumtion that the shape of the hierarchy pre-
and post-switch is similar enough.

In the end, I found that simply collecting all the "terminals" in this hierarchy
(the actual textures), calculating their global offset and then doing the
animations based off that was the easiest solution to reason about. I.e.
"flattening". I might change my mind later.

FURTHER RELATED SMALLER CHANGES:

* Drawing offsets is rounded to ints to ensure we draw on pixel boundaries.
* Texture backgrounds were made transparent rather than white; makes the
  animation more smooth
* Upgrade of OffsetBox to a class to support alpha-blending

SOME HACKS TO POINT OUT

i.e. things that we might want to reconsider "soonish"

* parallel hierarchy for Note (see notes at the top of clef_address)
* SExpr is given an extra field address, that is _only_ used when the SExpr is
  the result of Note.to_s_expr. Single-use field: probably an anti-pattern.
* SimpleScore is the 3rd Score implementation (see notes at simple_score.py)
* Offset to class conversion: we kept the tuple-indexing out of lazyness.
* We reconstruct the full (globally addressed) tree for each cursor-switch;
    quite wasteful

TO BE DONE SOON:

Certain things are not yet in this commit, should be done "soon" hopefully:

* support for click-while-animating selecting the thing as animated at the time of clicking
* if keeping the (history)cursor pointing at the same thing pre- and post-
  animate is possible, we should do that.
* viewports don't animate yet (leading to a jerky motion on the first frame of
  animation)
* This implementation is for history.py only, i.e. the ELS'18 rendering of
  Notes. animating of "in context history" (ic_history.py) is a nice TODO.