Factor out "in context history" cursor handling
Up to this commit, the in-context-history widget was using a strict
subset of the regular history widget's clef/structure/construct
combination (the subset consists of: only cursor-setting)

In preparation of further divergence between history.py and
ic_history.py, e.g. in terms of how a cursor is expressed in the first
place, I've refactored an entire clef/structure/construct combination
out.