IC History's cursor as a sinle index in items
Rather than: a full s-address in the node tree. In other words: you can select
a single line, representing a Note, rather than any arbitrary element, in this
widget now.

Combined with the following refactoring:

bottom_up_construct => node_cata: this makes it more clear what pattern is
being followed (and idea in bottom_up_construct, of rolling a "decoration with
the current s_address" into a catamorphism is no longer required, because the
cursor is not expressed )